+++ frontend/src/components/common/smart/dashboard/Dashboard.tsx
@@ -1,6 +1,6 @@
 import Layout from "@src/components/shared/layout";
-import DashboardSkeleton from "@src/components/shared/loaders/dashboardSkeleton/DashboardSkeleton";
-import RowSkeletonCard from "@src/components/shared/loaders/rowSkeletonCard/RowSkeletonCard";
+import DashboardSkeleton from "@src/components/shared/loaders/dashboardSkeleton/dashboardSkeleton";
+import RowSkeletonCard from "@src/components/shared/loaders/rowSkeletonCard/rowSkeletonCard";
 
 import { useGetEmployeesQuery } from "@src/store/reducers/employees-api";
 
@@ -13,13 +13,11 @@ const Dashboard = () => {
       ) : (
         <>
           <DashboardSkeleton />
+          <RowSkeletonCard />
         </>
       )}
       {/* To Test Display Skeleton! */}
       {/* <DashboardSkeleton /> */}
-      {new Array(20).fill(0).map((_, index) => (
-        <RowSkeletonCard key={index} />
-      ))}
     </Layout>
   );
 };
