--- frontend/src/components/shared/loaders/rowSkeletonCard/RowSkeletonCard.tsx
@@ -1,22 +1,0 @@
-import Box from "@mui/material/Box";
-import Skeleton from "@mui/material/Skeleton";
-import "./rowSkeletonCard.scss";
-
-function RowSkeletonCard() {
-  return (
-    <div className="row-skeleton">
-      <Box sx={{ width: "100%" }}>
-        <Skeleton
-          animation="wave"
-          height={90}
-          sx={{
-            backgroundColor: "#FFCDCD",
-            opacity: 0.2,
-            borderRadius: "12px",
-          }}
-        />
-      </Box>
-    </div>
-  );
-}
-export default RowSkeletonCard;
